requirements.txt
discord.py==2.0.1
Flask==1.1.2
pandas==1.3.4
requests==2.26.0
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.3